+++ tests/e2e/app.spec.js
@@ -72,7 +72,7 @@ test( `shows auth immediately for anonymous visitors`, async ( { page } ) => {
 
     await page.goto( `/` )
 
-    await expect( page.getByRole( `heading`, { name: `Sandbox Grapevine` } ) ).toBeVisible()
+    await expect( page.getByRole( `heading`, { name: `Sandbox, Grapevine` } ) ).toBeVisible()
     await expect( page.getByRole( `button`, { name: `Signup` } ) ).toBeVisible()
 } )